Skip traversing unimportant connections
A connection from a client to host process only needs to be evaluated if
the client is more important than the host is some dimension. Otherwise,
the connection can be skipped.

Also, if none of the targets of an update has a reduction in importance,
the reachable processes do not need to be recalculated before traversal
because their importance values can only stay the same or increase.

(Enable/disable with adb shell device_config put backstage_power
com.android.server.am.skip_unimportant_connections true/false)

Bug: 323376416
Test: atest MockingOomAdjusterTests

Change-Id: I1b7dcad4bb20036e41e0dbc88c002540600afb93

Former-commit-id: 0160e5ccf912ed74cd19402d556f959c2a9b2136