Separate RequestProcessor from interface
This is in preparation for a new implementation to add
support for private profile handling. The interactions
between work and private profile combine in ways that
make it simpler to create  new version with the new
requirements in mind.

The implementation will be switched via Dagger binding,
controlled by a flag.

Bug: 327613051
Test: N/A; pure mechanical refactor, no changes
Flag: N/A
Change-Id: I4ea124fbaa22c1afe1ce999016f4976e43598dfa

Former-commit-id: 37a10c94bf29783564a90a9c82b2effcd7db8863