Don't disable isHomeLaunchDelayable on HSUM
Headless System User Mode (HSUM) also requires theming to be complete
before launching SetupWizard (SUW). So we treat HSUM the same as
non-HSUM in this regard.

We continue to not enable it on Android Auto.

It was previously disabled on HSUM due to causing Auto's system to hang,
but hopefully since the current UserId is now evaluated correctly, this
is no longer necessary. Nonetheless, Auto hits the timeout, so it isn't
solving the problem and is adding a long boot delay. We therefore
continue to disable it on Auto.

Bug: 283904547
Test: Manually confirmed that SUW had theming on non-Auto HSUM device
Test: device_boot_test_aae_toolbox
Change-Id: I37242cdaae62582f0970a3a63c111e88096d66df

Former-commit-id: 6f65c3afa4c3c901f02b3a1e749ae8092d3bc159