Add a key to provide the default capabilities for satellite service
In case the entitlement server provides PLMN list and they are not in the list of carrier config, capabilities for those PLMNs will be empty.
Therefore it is needed to provide a way to set the default value for that scenario.
KEY_SATELLITE_DEFAULT_SERVICES_FOR_ENTITLEMENT_PLMNS_INT_ARRAY contains
values sms, mms and emergency is added.

Bug: 332811510
Test: Manually run below scenario and verified with the logs.
  1. conected to mock satellite service and run test entitlement server app
  2. override carrier config for newly added key
  3. turn on airplane mode and off
  4. verify whether entitlement provided plmns are updated with the default capabilitys.
  5. verify whether KEY_CARRIER_SUPPORTED_SATELLITE_SERVICES_PER_PROVIDER_BUNDLE provided PLMNs are not affectec by default capabilities.
Change-Id: I247121447c7356497233ba1ca230c5b8a9fe0f66

Former-commit-id: 5f76b93583bf412bc6b4e4c68da073b68753a402