[flexiglass] CentralSurfaces is a dep of StatusBarStateController.
This is needed to fix the crash in the attached bug.

The problem is that "root" is accessed before it's set.

The SysuiStatusBarStateController.start codepath accesses it.
The CentralSurfacesImpl.start sets it.

By adding the dependency, we ask the system to guarantee that
CentralSurfacesImpl.start happened before
SysuiStatusBarStateController.start is invoked.

Fix: 332956676
Test: manually verified that the crash no longer occurs and system UI
shows up normally (it was a crashloop)
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT

Change-Id: I15b679616e7b98d6c142fba55e891ee351fee997

Former-commit-id: 9a2e20a9f582373c972446253d08cdc39f7afd74