Use existing screenshot if present when type is FULLSCREEN
When ScreenshotController receives a ScrenshotData with
TAKE_SCREENSHOT_FULL_SCREEN, but a non-null Bitmap, treat
it the same as if it were TAKE_SCREENSHOT_PROVIDED_IMAGE.

This allows ScreenshotRequestProcessor to handle all cases,
while not affecting the accuracy of logging.

Rationale:

ScreenshotRequestProcessor is responsible for populating
the top component name, but defers to ScreenshotController
to actually capture the screen in full-screen cases.

With this change (and new PolicyRequestProcessor) the screen
will always be captured before moving to ScreenshotController.

This change is safe unflagged because this particular condition
does not occur currently. It will be a state provided when
the flag 'screenshot_private_profile' is enabled.

Bug: 327613051
Test: atest TakeSCreenshotExecutorTest TakeScreenshotServiceTest
Flag: N/A
Change-Id: If32a01f04d3cf0015817297db4c06a96da8057e8

Former-commit-id: eba137145819e87717278e9036e07b0c26b1aa17