Always use system config insets types for insets frame
When the half-way insets configuration decoupled solution was
introduced, an arg to indicate that the legacy value should be used was
introduced. This is no longer needed as the default value will always be
fully decoupled or coupled. When the default value is fully decoupled,
the method will not be called when calculating orientation, where we
need the legacy value.

Remove the arg is a choice and can reduce the complexity of that method.

Test: Check outStableBounds value
Test: DisplayPolicyInsetsTests
Test: DisplayPolicyTests
Bug: 327313645
Change-Id: Ia7cc14b5ee147fb3fedc70282e4cfb10eba3c92e

Former-commit-id: b2fecc4e964050aac62752143c2752ff9e71cbcd