Start using InputMethodSettingsRepository (2nd try)
This is the 2nd attempt of submitting my previous CL [1], which got
reverted due to a boottime regression (Bug 329414077). We have already
addressed the major contributor of the boot time regression [3], so
this CL should be ready to be submitted again.

With this CL we start taking advantage of

  InputMethodSettingsRepository,

which enables us to get a cached InputMethodSettings instead of
dynamically querying available IMEs.

To keep the CL as minimum as possible,

  InputMethodManagerService#mSettings

remains unchanged. It will be removed in a subsequent bug 329163064.

There must be no user observable behavior change.

 [1]: Ie2d3b7e9b75790f64789d0c2e96ae46b1dee3a1e
      c305c29fa5558e1a7ec34d2d063bb150ec95244d [formerly 03e0580e6f1ff8c4bb345747448b2569ebf7f1bf]
 [2]: I20317bdd012c51a5ba57cd6e4d2e1a69df66d6a2
      5d4a811131771898526ceb65c10c02463cde320e [formerly 84c61c453fdceb0eb48d05c4ea5620bbcd51f952]
 [3]: I8e51b1ced4dc16cdca7e898885c64793665fafef
      26d1612bcc29250756cb40f50b76e32dacc11936 [formerly bce89b5742781e680efac88cbc7179bf029638d9]

Bug: 329163064
Bug: 333591545
Fix: 309837937
Test: presubmit
Change-Id: Ic0dd655fbd86b8ccce2b3298b4c70359a468f9ec

Former-commit-id: 8130073e978ab0a826ecfab9bee1254551023a32