Revert^2 "Correct activity's lifecycle when the process was killed in background."
The lifecycle of an invisible activity would stay in pause stage if
it's process was died and restore from attachApplication. The incorrect
status can affect transition since it won't be removed from
UnknownAppVisibilityController. Stopping the invisible activity so it
won't stay in paused.

Bug: 327596503
Test: atest RecentsAnimationTest
Test: launch a translucent app, entering keyguard, kill home process.
Verify the activity of home process would be stopped after attach the
process back.

Change-Id: I7560c40ddf1a223e2a2aaba463124d06cc3c0051
Merged-In: I80756609fe643877971408934b19ad97aa0ae746

Former-commit-id: b7c1ba2dfa699c583604765160bff36cb7ee3b76