Removes posting to a handler for onWindowInfosChanged.
This handler post was originally added to fix an ANR in b/221497060 but
that underlying lock issue has since been resolved in b/222421815.

Now, this handler post can cause OOM in extreme cases when hundreds of
messages are posted before being handled.
It should be safe to remove given that the locking issue is resolved.

Bug: 333834990
Test: Drag app from taskbar to split screen, observe no ANR
Change-Id: I5e391bb375105f369586374e38a8980a12b449b8

Former-commit-id: cfd152e5d9db83023148c2d9b0d7ae21473f8921