SpatializerHelper: assimilate speaker safe to speaker device for SA compatibility
Although speaker safe is a mono device and should not be compatible with
transaural spatilization, we consider it is to avoid transient SA availability
changes propagated to apps that would cause failures to play SA content.
This can happen for instance if ringtone playback is still active when
the onRoutingUpdated() calback is called by APM due to an audio mode change.

A better longer term solution would be to trigger the onRoutingUpdated()
callback more dynamically each time the media routing changes but the
impact on performance and other consumers of this callback must be investigated.

Bug: 331349123
Test: repro steps in the bug
Change-Id: Id5d48427ff4ecef02a01a6e7fe46c5f910825e81

Former-commit-id: e7b122acd4ad7f2819e3d106dd85c9ffc8a77fac