Remove android.security.KeyStore
Since android.security.KeyStore is no longer used, remove it.  This
prevents this class, which had a name that made it sound much more
important than it actually is, from being confused with the classes via
which the Android Keystore is actually accessed these days.

Bug: 326508120
Test: build
Flag: NONE mechanical refactoring with no behavior change
Change-Id: I2f8471971b2816a4a830a48cb99ff118c21a6ad8

Former-commit-id: a8516526e2028255ab8c3c52688c5b4cb4be58da