Allow creating ASurfaceControl from NativeActivity Surface
When generating a Surface, we have the option to include the
associated SurfaceControl handle. This handle is necessary
for creating child ASurfaceControl parented to a ANW.
Previously, the SurfaceControl handle was only included in
the SurfaceView surface, but not in the viewrootimpl surface.
This meant NativeActivity users could not attach child
surfaces. Fix this by including the surfacecontrol handle
into the VRI surface.

Bug: 320706287
Test: atest FrameworksCoreTests:ViewRootImplTest
Change-Id: I0af8c1b53f616f2ec5fb4c927553d197d1192daa

Former-commit-id: 34459252775f3d09f36000e897b80c76c6ee7adb