Reapply "Finalize resources for VIC"
This reverts commit 45f5c3f48ccf8c4606c2070c5434c73794a7bdf0 [formerly 111380a17601bf006b4cfee975e4ff6c137782da].

Instead of listing the resources twice, just mark the original ones as
removed. This will break any existing clients of the APIs until they
recompile, but given that flags weren't accesible to application for
most of this release (and still aren't available to most apps), this
should have minimal-to-no impact in practice.

Bug: 335304236
Test: atest --host  WearSdkRoboTests:WearSdkVersionTest
Change-Id: I7998a1f28e51e40e911ce6d51412232714993a84

Former-commit-id: 7fa5818fb7037c541807c3754882d98f6ea22d20