Add partial TAPL support for Scene Container
This is the first in a series of CLs that will make Platform Scenario
Tests pass with the scene container flag enabled. The approach is to
add test tags to composables and update TAPL lookups to use those IDs
when the scene container flag is enabled.

Bug: 338267587
Test: made NotificationRemoteInput.remoteInputUpdate pass locally
Flag: ACONFIG com.android.systemui.scene_container DEVELOPMENT
Change-Id: I0e9360f16932730375374cea6505ca192dd2dea2

Former-commit-id: d13d43543b0ff35e2658017c84f575e971401297