Move SideFpsController requests add call
Only add a show request to the request set if the request goes through
(overlayView == null). Adding the show request even for ignored requests
may be the source of the indicator persisting on screen when it should
be hidden, because the hide() method checks if the set is empty before
setting the overlayView to null.

Flag: NONE
Fixes: 338778802
Test: None (cannot reproduce reported issue)
Change-Id: Ic50c0709f4c62385a2aa941f7b3e1221e663a1d2
Merged-In: Ied1e743db40ecffc0ed07aa506eaf9be82c57a0a

Former-commit-id: e831fce01755cae8523fbe2c42569816bea15bb0