Make tf organizer process have visible state with visibile activity
This is similar to [1]. But even if the activity is not embedded, the
organizer process will still need to response for task fragment info
change. So if there is any visible activity on a task which is
associated with tf organizer, the organizer process also need to be
visible state to handle the event.

[1]: I57ef9ebe95e05842e7f325c1a4775e98c7eb57bd

Fix: 337784723
Test: atest TaskFragmentOrganizerControllerTest# \
            testOnActivityReparentedToTask_ \
            activityNotInOrganizerProcess_useTemporaryToken

Change-Id: I42c9cb2b95c5517370ef2d26b948b3f4e0db8ec4

Former-commit-id: 65b9bff7eb4a31ea3e47d862489a1dfb873a55a9