Add a new feature flag in android.net.wifi
Add a new feature flag - "network_provider_battery_charging_status" in
android.net.wifi package.
It is the same flag which was added in ag/25092320 in the wifi module.

framework-minus-apex depends on a flag defined in wifi_aconfig_flags
but framework-minus-apex does not depend on the corresponding aconfig
java lib wifi_aconfig_flags_lib. Therefore Soong doesn't know about
the dependency -> Soong won't tell Metalava about the flag ->
Metalava will default to reverting the API to the previous SDK
snapshot -> API will not be present in the SDK and docs artifacts.

Solution:
1) define a new flag in frameworks/base
2) Work with relteam to ramp it to next ASAP
3) Move NetworkProviderInfo to use this new flag

Bug: 339181644
Test: N/A
Change-Id: I358638546dc888a41498ec066d548c6612ea4f60

Former-commit-id: 28493b06879c02415dbd1c2ea77ba412c5e11dfc