Handle different font scale values during B&R
The DEFAULT_DEVICE_FONT_SCALE System Settings should not be
transferred during a D2D session because read directly from
a configuration resource.

When we restore a device we need to check if the current font_scale
is allowed. If not we update to the first allowed value which is
equals or larger than the current one

Bug: 337988698
Test: atest SettingsProviderTest:SettingsBackupTest
Test: atest SettingsProviderTest:SettingsBackupAgentTest

Change-Id: Ida0bb54e34de472c9eabf37a8ae6fb2aef0ee541

Former-commit-id: f577dbae970aa17cc555154901cba9cd890183a8