Allow for resize into the same bounds
In PiP2, resizes are transitions, so we need
to make sure we do not schedule a no-op transition
that can be aborted. Hence we do the following:

For fling gestures: do not schedule a transition if
final motion bounds are invariant from the original bounds.

For pinch-resizing: if the final bounds are unvariant,
in order to have the animation after release
running as a part of a transition, use a trick where we offset
the destination bounds by 1 pixel away from the closes edge.

Also enable dump() callbacks in PiP2, for easier debugging and
state checks. Currently PipController and PipTouchHandler are
the entry points (since they are independent in PiP2).

Bug: 340883807
Test: atest WMShellUnitTests:PipTransitionStateTest
Test: enter PiP2, pinch resize to min bounds, and drag beyond bottom
right corner

Change-Id: I0ec15644d345bd8ca25aa42259ca938360ab54ad

Former-commit-id: 3ebff91826bb818a55e62f5ca48c368b1141f0e6