Avoid dismissing shade when launching activities from communal hub
As the shade is closed when an activity is launched from the hub, trying
to dismiss the shade will cause it to flicker open. This change avoids
this by not dismissing the shade.

This change also ensures that the hub remains visible while a
showWhenLocked activity opens over the hub.

Bug: 331591373
Fixes: 342395651
Flag: com.android.systemui.glanceable_hub
Test: atest SystemUiRoboTests
Change-Id: I550b646a1a1572a062d9a4c8090cd16dd60650aa

Former-commit-id: 5df44e990f5d8e50ec246bf6f7b4fdab4a150ecb