Always reset interruption deltas
This CL ensures that we always reset the interruption deltas when an
interruption happens. This is usually not necessary, but in some
specific timings (e.g. if a transition is finished right after an
interruption occured) not cleaning might lead the unexpected
interruption deltas being animated.

Bug: 290930950
Test: ElementTest
Flag: com.android.systemui.scene_container
Change-Id: I3e27690a10270ee71d54677f77bca13f11002bc0

Former-commit-id: edc181e0081f4d34fca4c77f2c42307c9bf33c85