Binder: avoid GC while holding lock
The JavaBBinder constructor may trigger a GC. This
causes deadlocks.

Though, I also think, usually when you create a
binder object, you just send it out, and so maybe
we should always create the local object instead
of doing it lazily like this, but idk the history
behind that, I've been thinking of changing it.

Bug: 360067751
Test: boot, TH
Google: 3237098
Change-Id: I97963c5e7b859d4ad474c715bb657b9baa19fabf